Switching to time-based dosing (for the moment).
When doing initial experiments with the assembled project, the weight
measurements were found to be going a little crazy the moment the motor
turned on.  I'm not at the moment sure whether the problem is EMI from
the motor, or if it's just vibration.  In any case, the load cell method
is non-viable for the moment, and I need my coffee grinder.  So:
time-based measurements it is for now.